% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BivMultinominal}
\alias{BivMultinominal}
\title{Generate Bivariate Multinomial Categorical Data}
\usage{
BivMultinominal(N, zeta = 1.5)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{zeta}{Numeric. Strength of interaction. Default is 1.5.}
}
\value{
A data frame with columns Z1, Z2, X, and Y (both factors).
}
\description{
Creates a multinomial dataset where the probabilities are nonlinear functions of Z1 and Z2.
}
