% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BivNonLinearCategorization}
\alias{BivNonLinearCategorization}
\title{Generate Bivariate Nonlinear Categorical Data}
\usage{
BivNonLinearCategorization(N, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Generates categorical variables X and Y based on nonlinear combinations of Z1 and Z2.
}
