% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cci.R
\name{calculate_cci}
\alias{calculate_cci}
\title{Calculate CCI}
\usage{
calculate_cci(
  x,
  id = "id",
  code = "code",
  cci = "cci",
  keep_groups = FALSE,
  codelist = cci_codelist_icd8_icd10,
  sks_codes = TRUE
)
}
\arguments{
\item{x}{data.frame}

\item{id}{Name of id variable in \code{x}}

\item{code}{Name of variable with codes in \code{x}}

\item{cci}{Name of CCI variable in return data.table}

\item{keep_groups}{Logical. Keep individual CCI disease group variables in
return data.frame?}

\item{codelist}{Codelist with CCI disease gorup definitions and weightds used
to define the CCI. By default the \code{cci_codelist_icd8_icd10} codelist that
comes with package is used. This is reasonable codelist to use if ´x´ is
data from the Danish National Patient Registry. Alternatively, a custom
codelist can be provided.}

\item{sks_codes}{Logical. Convert SKS codes to ICD-10 in data? By default it
is assumed that the function is used on Danish registry data where ICD-8 and
ICD-10 codes are mixed, and ICD-10 codes have a "D" prefix.}
}
\value{
data.table if \code{x} data.table. Else, a data.frame
}
\description{
Calculate CCI
}
\examples{
set.seed(1)
dat <- data.frame(id = rep(1:10, each = 10), code = sample_codes(100))
calculate_cci(dat)
}
