% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci.R
\name{cci}
\alias{cci}
\title{Cyclic Causal Inference (CCI) algorithm.}
\arguments{
\item{suffStat}{list of sufficient statistics needed by the CI test. E.g., the data or covariance matrix}

\item{indepTest}{the CI test function}

\item{alpha}{the alpha value for the CI test}

\item{p}{the number of variables}

\item{skeleton_pre}{Pre-computed skeleton}

\item{rules}{A logical vector indicating which of the 7 orientation rules to fire}
}
\value{
A list containing the partially oriented MAAG \code{paag} and statistic \code{Sta}
}
\description{
Discovers a partially oriented maximal almost ancestral graph (MAAG) of a directed graph G,
provided that the global Markov property and d-separation faithfulness holds according to G.
}
