% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCI.tuner.R
\name{CCI.tuner}
\alias{CCI.tuner}
\alias{tuner}
\title{CCI tuner function for CCI test}
\usage{
CCI.tuner(
  formula,
  data,
  method = "rf",
  random_grid = TRUE,
  metric = "RMSE",
  folds = 5,
  metricfunc = NULL,
  mlfunc = NULL,
  ml_param = NULL,
  ...
)
}
\arguments{
\item{formula}{Model formula or a DAGitty object specifying the relationship between dependent and independent variables.}

\item{data}{A data frame containing the variables specified in the formula.}

\item{method}{Character. Specifies the machine learning method to use. Supported methods include generlaized linear models "lm", random forest "rf", and extreme gradient boosting "xgboost", etc. Default is "rf".}

\item{random_grid}{Logical. If TRUE, a random grid search is performed. If FALSE, a full grid search is performed. Default is TRUE.}

\item{metric}{Character. Specifies the performance metric to optimize during the tuning process. Default is "RMSE".}

\item{folds}{Integer. The number of folds for cross-validation during the tuning process. Default is 5.}

\item{metricfunc}{Optional custom function for calculating a performance metric based on the model's predictions. Default is NULL.}

\item{mlfunc}{Optional custom machine learning function to use instead of the predefined methods. Default is NULL.}

\item{ml_param}{List. A list of parameters to be tuned when using the mlfunc. Default is NULL.}

\item{...}{Additional arguments to pass to the \code{CCI.tuner} function.}
}
\value{
Returns tuned parameters values for the predictive function
}
\description{
The \code{CCI.tuner} function performs a grid search over parameters for a conditional independence test using a specified machine learning model or a custom model provided by the user. It calculates the test statistic, generates a null distribution via permutations, computes p-values, and optionally generates a plot of the null distribution with the observed test statistic.
}
\seealso{
\code{\link{CCI.test} \link{perm.test}}, \code{\link{print.summary.CCI}}, \code{\link{plot.CCI}}, \code{\link{QQplot}}
}
