% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cci_codelist_icd10}
\alias{cci_codelist_icd10}
\title{Codelist mapping ICD10 codes to Charlson Comorbidity
Index disease groups}
\format{
A list with 19 elements, where each element is also a list
with 3 elements used to define each disease group:
\describe{
\item{label}{Disease group label}
\item{name}{Name used as variable name in data}
\item{codes}{Character vector of ICD-10 codes with no punctuation.}
\item{assign0_if}{Character vector with names of CCI disease groups that is
used to make corrections to this disease group, eg if one of the listed
disease groups are present, the current disease group will be defined as
being absent, regardless of whether or not it actually is.}
}
}
\usage{
cci_codelist_icd10
}
\description{
A list containing the labels, names and codes used to define
each of the disease groups in the Charlson Comorbidity Index using
ICD-10 codes.
}
\keyword{datasets}
