% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ComplexCategorization}
\alias{ComplexCategorization}
\title{Generate Complex Categorical Data}
\usage{
ComplexCategorization(N)
}
\arguments{
\item{N}{Integer. Sample size.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
A more intricate categorization based on combinations of Z1 and Z2.
}
\examples{
head(ComplexCategorization(100))

}
