% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ExpLogData}
\alias{ExpLogData}
\title{Generate Categorical Data Based on Exponential and Logarithmic Functions}
\usage{
ExpLogData(N, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{d}{Numeric. Dependency strength. Default is 0.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Categorizes based on thresholds of exponential and logarithmic transformations of Z1 and Z2.
}
