% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{expLogThresholdContXSim}
\alias{expLogThresholdContXSim}
\title{Generate Exponential and Logarithmic Data}
\usage{
expLogThresholdContXSim(N)
}
\arguments{
\item{N}{Integer. Sample size.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Generates data with exponential and logarithmic dependencies based on Z1 and Z2.
}
\examples{
head(expLogThresholdContXSim(100))
}
