% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tuned_params}
\alias{get_tuned_params}
\title{Get the best parameters after tuning with CCI.tuner}
\usage{
get_tuned_params(tuned_model)
}
\arguments{
\item{tuned_model}{A model object returned from the CCI.pretuner function. This object contains the tuned parameters and other relevant information.}
}
\value{
A named list of tuned parameters specific to the model method (e.g., \code{mtry} for random forest, \code{eta}, \code{max_depth} for xgboost), named as the arguments of \code{\link{CCI.test}}. Returns \code{NULL} for unsupported methods.
The xgboost row \code{subsample} is not returned, since \code{subsample} in \code{CCI.test} controls subsampling of the data.
}
\description{
Get the best parameters after tuning with CCI.tuner
}
