% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gridPartitionContXSim}
\alias{gridPartitionContXSim}
\title{Generate Grid Partitioned Data}
\usage{
gridPartitionContXSim(N)
}
\arguments{
\item{N}{Integer. Sample size.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Generates data with a grid partitioning effect based on Z1 and Z2.
}
\examples{
head(gridPartitionContXSim(100))

}
