% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hard_case_sim}
\alias{hard_case_sim}
\title{Generate Hard Case Data}
\usage{
hard_case_sim(N)
}
\arguments{
\item{N}{Integer. Sample size.}
}
\value{
A data frame with columns X, Y, and Z.
}
\description{
Generates data with a hard case scenario where X and Y are influenced by Z in a nonlinear manner.
}
\examples{
head(hard_case_sim(100))
}
