% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HardCase}
\alias{HardCase}
\title{Example dataset: HardCase}
\format{
A data frame with 500 rows and 3 variables:
\describe{
\item{X}{Numeric vector}
\item{Y}{Numeric vector}
\item{Z1}{Conditioning variable}
\item{Z2}{Conditioning variable}
}
}
\source{
Simulated data.
}
\usage{
HardCase
}
\description{
A dataset containing simulated conditional independence test results.
}
\keyword{datasets}
