% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10_cm_2020}
\alias{icd10_cm_2020}
\title{ICD-10-CM codes, 2020 version}
\format{
A data.frame with 72184 rows and 2 columns containing
ICD-10-CM codes.
\describe{
\item{code}{ICD-10-CM code}
\item{label}{Code label}
}
}
\source{
\url{https://ftp.cdc.gov/pub/Health_Statistics/NCHS/Publications/ICD10CM/2020/icd10cm_codes_2020.txt}
}
\usage{
icd10_cm_2020
}
\description{
ICD-10-CM, 2020 version codes with corresponding labels.
}
\keyword{datasets}
