% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd8_4digits}
\alias{icd8_4digits}
\title{ICD-8 codes}
\format{
A data.frame with 4510 rows and 3 columns containing
ICD-8 codes. Note that not all subcodes are included.
\describe{
\item{code}{ICD-8 code}
\item{chapter}{Chapter label}
\item{label}{Code label}
}
}
\source{
\url{http://www.wolfbane.com/icd/icd8h.htm}
}
\usage{
icd8_4digits
}
\description{
ICD-8 codes and corresponding chapter(labels). Note that
not all sub codes are included.
}
\keyword{datasets}
