% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{InteractiondData}
\alias{InteractiondData}
\title{Generate Categorical Data Based on Interactions}
\usage{
InteractiondData(N, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{d}{Numeric. Dependency strength. Default is 0.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Creates categorical X and Y variables based on the interaction of signs and sums of Z1 and Z2.
}
