% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IP_discovery.R
\name{IP_discovery}
\alias{IP_discovery}
\title{FCI's skeleton discovery procedure adapted for CCI}
\usage{
IP_discovery(
  suffStat,
  indepTest,
  alpha,
  p,
  labels = labels,
  m.max = Inf,
  fixedGaps = fixedGaps,
  fixedEdges = fixedEdges,
  method = c("stable", "original", "stable.fast"),
  NAdelete = TRUE,
  numCores = 1,
  verbose = TRUE
)
}
\arguments{
\item{suffStat}{List containing suffStat$data}

\item{indepTest}{The independence test}

\item{alpha}{Type I error rate for indepTest}
}
\value{
List containing the skeleton data
}
\description{
FCI's skeleton discovery procedure adapted for CCI
}
