% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_categorical_Z_any}
\alias{is_categorical_Z_any}
\title{Check whether Z contains at least one categorical variable}
\usage{
is_categorical_Z_any(sub_data, Z, allow_character = TRUE)
}
\arguments{
\item{sub_data}{data.frame containing the Z columns.}

\item{Z}{character vector of column names defining the conditioning set.}

\item{allow_character}{logical; treat character as categorical. Default TRUE.}
}
\value{
logical scalar.
}
\description{
Categorical is defined as factor (and optionally character).
}
