% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{multi_class_log_loss}
\alias{multi_class_log_loss}
\title{Calculation of log loss for categorical outcome}
\usage{
multi_class_log_loss(actual, predicted, all_levels)
}
\arguments{
\item{actual}{The observed categorical (factor) outcome variable.}

\item{predicted}{The predicted probabilities for each category.}

\item{all_levels}{A vector of all possible levels (categories) of the outcome variable.}
}
\value{
Log loss of classification model
}
\description{
Calculation of log loss for categorical outcome
}
