% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Multinominal}
\alias{Multinominal}
\title{Generate Multinomial Categorical Data}
\usage{
Multinominal(N, zeta = 1.5)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{zeta}{Numeric. Strength of interaction. Default is 1.5.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Multinomial categorical variables X and Y generated via nonlinear logits of Z1 and Z2.
}
