% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{NonLinearCategorization}
\alias{NonLinearCategorization}
\title{Example dataset: NonLinearCategorization}
\format{
A data frame with 600 rows and 3 variables:
\describe{
\item{X}{Numeric vector}
\item{Y}{Numeric vector}
\item{Z}{Conditioning variable}
}
}
\source{
Simulated data.
}
\usage{
NonLinearCategorization
}
\description{
A dataset containing simulated data from a non-linear transformation followed by categorization.
}
\keyword{datasets}
