% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{NonLinNormal}
\alias{NonLinNormal}
\title{Generate Nonlinear Normal Data}
\usage{
NonLinNormal(N, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{d}{Numeric. Dependency strength. Default is 0.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Creates nonlinear continuous data based on an exponential interaction of Z1 and Z2.
}
