% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NonLinNormalZs_d0}
\alias{NonLinNormalZs_d0}
\title{Example dataset: NonLinNormalZs_d0}
\format{
A data frame with 1000 rows and 15 variables:
\describe{
\item{X}{Numeric vector}
\item{Y}{Numeric vector}
\item{Z1}{Conditioning variable}
\item{Z2}{Conditioning variable}
\item{Z3}{Conditioning variable}
\item{Z4}{Conditioning variable}
\item{Z5}{Conditioning variable}
\item{Z6}{Conditioning variable}
\item{Z7}{Conditioning variable}
\item{Z8}{Conditioning variable}
\item{Z9}{Conditioning variable}
\item{Z10}{Conditioning variable}
\item{Z11}{Conditioning variable}
\item{Z12}{Conditioning variable}
\item{Z13}{Conditioning variable}
\item{Z14}{Conditioning variable}
\item{Z15}{Conditioning variable}
}
}
\source{
Simulated data.
}
\usage{
NonLinNormalZs_d0
}
\description{
A dataset containing simulated data with uniform noise.
}
\keyword{datasets}
