% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{NormalData}
\alias{NormalData}
\title{Generate Normal Data for Conditional Independence Testing}
\usage{
NormalData(N)
}
\arguments{
\item{N}{Integer. Sample size.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
This function generates continuous data where X and Y are both functions of Z1 and Z2 with added normal noise.
}
