% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PoissonNoise}
\alias{PoissonNoise}
\title{Generate Data with Poisson Noise}
\usage{
PoissonNoise(N, lambda = 1)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{lambda}{Numeric. Rate parameter for the Poisson distribution. Default is 1.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Adds Poisson noise to a nonlinear combination of Z1 and Z2.
}
\examples{
head(PoissonNoise(100))

}
