% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PolyData}
\alias{PolyData}
\title{Example dataset: PolyData}
\format{
A data frame with 600 rows and 4 variables:
\describe{
\item{X}{Numeric vector}
\item{Y}{Numeric vector}
\item{Z1}{Conditioning variable}
\item{Z2}{Conditioning variable}
}
}
\source{
Simulated data.
}
\usage{
PolyData
}
\description{
A dataset containing simulated data from a polynomial relationship.
}
\keyword{datasets}
