% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.codelist.R
\name{print.codelist}
\alias{print.codelist}
\title{Print method for codelists}
\usage{
\method{print}{codelist}(x, ...)
}
\arguments{
\item{x}{a codelist}

\item{...}{Other arguments}
}
\description{
Print method for codelists
}
\examples{
print(cci_codelist_icd10)
}
