% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{protein_data}
\alias{protein_data}
\title{Protein signaling dataset}
\format{
A data frame with 7466 rows and 11 variables:
\describe{
\item{PKC}{Protein kinase C}
\item{PKA}{Protein kinase A}
\item{p38}{Mitogen-activated protein kinase p38}
\item{JNK}{c-Jun N-terminal kinase}
\item{ERK}{Extracellular signal-regulated kinase}
\item{Akt}{Protein kinase B (Akt)}
\item{Raf}{Raf proto-oncogene serine/threonine-protein kinase}
\item{Mek}{Mitogen-activated protein kinase kinase}
\item{PIP2}{Phosphatidylinositol 4,5-bisphosphate}
\item{Plcg}{Phospholipase C gamma}
\item{PIP3}{Phosphatidylinositol (3,4,5)-trisphosphate}
}
}
\source{
\doi{10.1126/science.1105809}
}
\usage{
protein_data
}
\description{
A widely used dataset from flow cytometry experiments on human T-cells,
originally published by \insertCite{Sachs2005;textual}{CCI}.
The data records the expression levels of multiple proteins and phospholipids
in single cells, under different experimental interventions. It has been
frequently used as a benchmark dataset in the causal discovery literature.
}
\details{
\itemize{
\item Each row corresponds to a single cell measurement.
\item Values are continuous protein expression levels (fluorescence intensities).
\item The dataset has been used extensively to test causal discovery algorithms.
}
}
\keyword{datasets}
