% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{quadThreshContXSim}
\alias{quadThreshContXSim}
\title{Generate Quadratic Threshold Data}
\usage{
quadThreshContXSim(N)
}
\arguments{
\item{N}{Integer. Sample size.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Generates data with a quadratic threshold effect based on Z1 and Z2.
}
\examples{
head(quadThreshContXSim(100))
}
