% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_codes.R
\name{sample_codes}
\alias{sample_codes}
\title{Sample codes}
\usage{
sample_codes(
  n = 1L,
  code_sources = c("icd8_4digits", "icd10_cm_2020"),
  max_code_length = 4
)
}
\arguments{
\item{n}{positive integer. Number of codes to sample.}

\item{code_sources}{Character vector. Character vector with dataset names
from the package form which codes are sampled. By default codes are sampled
from the \code{icd8_4digits} and \code{icd10_cm_2020} dataset that is included in the
package. The complete list of datasets from which it is possible to sample
is: \code{icd8_4digits} and \code{icd10_cm_2020}.}

\item{max_code_length}{positive integer. Maximum length of codes to sample.
By default, the sampled codes are shortened to length 4.}
}
\value{
character vector of length \code{n}.
}
\description{
Sample ICD-8 and/or ICD-10 codes. The codes are sampled with replacement.
}
\examples{
sample_codes(5L)
}
