% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sinCosThresholdContXSim}
\alias{sinCosThresholdContXSim}
\title{Generate Sinusoidal and Cosine Data}
\usage{
sinCosThresholdContXSim(N)
}
\arguments{
\item{N}{Integer. Sample size.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Generates data with sinusoidal and cosine dependencies based on Z1 and Z2.
}
\examples{
head(sinCosThresholdContXSim(100))
}
