% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{SineGaussian}
\alias{SineGaussian}
\title{Generate Sine-Gaussian Data (Univariate)}
\usage{
SineGaussian(N, a = 1, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{a}{Numeric. Frequency parameter of the sine function. Default is 1.}

\item{d}{Numeric. Strength of dependency between X and Y. Default is 0.}
}
\value{
A data frame with columns Z, X, and Y.
}
\description{
This function generates data with a nonlinear sinusoidal dependency based on a Gaussian density envelope.
}
