% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sineGaussian_biv}
\alias{sineGaussian_biv}
\title{Generate Sine-Gaussian Data (Bivariate)}
\usage{
sineGaussian_biv(N, a = 1)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{a}{Numeric. Frequency parameter for the sine function. Default is 1.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
This function generates bivariate data with nonlinear dependencies based on a Gaussian density envelope and sinusoidal functions.
}
