% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sineGaussian_noise}
\alias{sineGaussian_noise}
\title{Generate Sine-Gaussian Data (Bivariate)}
\usage{
sineGaussian_noise(N, a = 1, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{a}{Numeric. Frequency parameter for the sine function. Default is 1.}

\item{d}{Numeric. Strength of dependency between X and Y. Default is 0.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
This function generates bivariate data with nonlinear dependencies based on a Gaussian density envelope and sinusoidal functions.
}
