% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{TrigData}
\alias{TrigData}
\title{Generate Categorical Trigonometric Data}
\usage{
TrigData(N, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{d}{Numeric. Dependency strength. Default is 0.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Uses sine and cosine functions of Z1 and Z2 to generate categorical outcomes.
}
