% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unclean_formula}
\alias{unclean_formula}
\title{Convert CI-style formula Y ~ X | Z into regression-style Y ~ X + Z}
\usage{
unclean_formula(formula)
}
\arguments{
\item{formula}{A formula of the form Y ~ X | Z1 + Z2, or already Y ~ X + Z1 + Z2.}
}
\value{
A standard formula of the form Y ~ X + Z1 + Z2.
}
\description{
Convert CI-style formula Y ~ X | Z into regression-style Y ~ X + Z
}
