% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{UniformNoise_large}
\alias{UniformNoise_large}
\title{Example dataset: UniformNoise_large}
\format{
A data frame with 20000 rows and 4 variables:
\describe{
\item{X}{Numeric vector}
\item{Y}{Numeric vector}
\item{Z1}{Conditioning variable}
\item{Z2}{Conditioning variable}
}
}
\source{
Simulated data.
}
\usage{
UniformNoise_large
}
\description{
A dataset containing simulated data with uniform noise.
}
\keyword{datasets}
