% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrapper_glm}
\alias{wrapper_glm}
\title{Wrapper function for GLM model training and evaluation}
\usage{
wrapper_glm(
  formula,
  data,
  train_indices,
  test_indices,
  family,
  data_type,
  metricfunc = NULL,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}

\item{train_indices}{Indices for training data}

\item{test_indices}{Indices for testing data}

\item{family}{Family for GLM}

\item{data_type}{Type of data (continuous or binary)}

\item{metricfunc}{A user-specified function which calculates a metric}

\item{...}{Additional arguments passed to glm}
}
\value{
Performance metric (defaults are RMSE for continuous, Kappa for binary)
}
\description{
Wrapper function for GLM model training and evaluation
}
