% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrapper_ranger}
\alias{wrapper_ranger}
\title{Random Forest wrapper for CCI}
\usage{
wrapper_ranger(
  formula,
  data,
  train_indices,
  test_indices,
  metric,
  metricfunc = NULL,
  nthread = 1,
  mtry = NULL,
  num.trees,
  eps = 1e-15,
  ...
)
}
\arguments{
\item{formula}{Model formula specifying the dependent and independent variables.}

\item{data}{Data frame containing the dataset to be used for training and testing the model.}

\item{train_indices}{A vector of indices specifying the rows in \code{data} to be used as the training set.}

\item{test_indices}{A vector of indices specifying the rows in \code{data} to be used as the test set.}

\item{metric}{Type of metric ("RMSE", "Kappa" or "Log Loss")}

\item{metricfunc}{Optional user-defined function \code{function(actual, predictions, ...)} returning a numeric performance value. For a factor response, \code{predictions} are the predicted classes. The \code{...} arguments are only passed on if the function accepts them.}

\item{nthread}{Integer. The number of threads to use for parallel processing. Default is 1.}

\item{mtry}{Integer. The number of variables to possibly split at in each node. Default is the square root of the number of columns in \code{data}.}

\item{num.trees}{Integer. The number of trees to grow in the random forest.}

\item{eps}{Small value to avoid log(0) in LogLoss calculations. Default is 1e-15.}

\item{...}{Additional arguments passed to the \code{ranger} function.}
}
\value{
A numeric value representing the performance metric of the model on the test set.
}
\description{
Random Forest wrapper for CCI
}
