% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{xgboost_wrapper}
\alias{xgboost_wrapper}
\title{Wrapper function for XGBoost model training and evaluation}
\usage{
xgboost_wrapper(
  formula,
  data,
  train_indices,
  test_indices,
  nrounds,
  objective,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}

\item{train_indices}{Indices for training data}

\item{test_indices}{Indices for training data}

\item{nrounds}{Number of boosting rounds}

\item{objective}{Objective function for XGBoost}

\item{...}{Additional arguments passed to xgb.train}

\item{iteration}{Current iteration index}
}
\value{
Performance metric (RMSE for continuous, Kappa for binary, Kappa for categorical)
}
\description{
Wrapper function for XGBoost model training and evaluation
}
