% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{BinaryData}
\alias{BinaryData}
\title{Generate Binary Data}
\usage{
BinaryData(N, threshold = 0, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{threshold}{Numeric. Threshold for binary classification. Default is 0.}

\item{d}{Numeric. Dependency strength. Default is 0.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Creates binary data based on a nonlinear interaction of Z1 and Z2.
}
\examples{
head(BinaryData(100))

}
