% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCI.direction.R
\name{CCI.direction}
\alias{CCI.direction}
\title{Choose Direction for testing for the CCI test}
\usage{
CCI.direction(
  formula,
  data,
  method = "rf",
  folds = 4,
  subsample = 1,
  poly = TRUE,
  degree = 3,
  interaction = TRUE,
  ...
)
}
\arguments{
\item{formula}{A formula object specifying the model to be fitted.}

\item{data}{A data frame containing the variables specified in the formula.}

\item{method}{A character string specifying the method to be used for model fitting. Options include "rf" (random forest), "xgboost" (XGBoost), "nnet" (neural network), "gpr" (Gaussian process regression), and "svm" (support vector machine).}

\item{folds}{An integer specifying the number of folds for cross-validation. Default is 4.}

\item{subsample}{Numeric. The proportion of the data to be used. Default is 1 (i.e., use all data).}

\item{poly}{Logical. If TRUE, polynomial terms of the conditioning variables are included in the model. Default is TRUE.}

\item{degree}{Integer. The degree of polynomial terms to include if \code{poly} is TRUE. Default is 3.}

\item{interaction}{Logical. If TRUE, interaction terms of the conditioning variables are included in the model. Default is TRUE.}

\item{...}{Additional arguments to be passed to the model fitting function.}
}
\value{
A formula object specifying the selected model direction.
}
\description{
This function selects the best direction for the CCI test based on cross validation. For the condition Y \emph{||} X | Z, the function return the recommended
formula either Y ~ X | Z or X ~ Y | Z .
}
\examples{
# Example usage
if (requireNamespace("randomForest", quietly = TRUE)) {
  dat <- SineGaussianBiv(N = 100, a = 2)
  formula_init <- Y ~ X + Z1 + Z2
  chosen_formula <- CCI.direction(
    formula = formula_init,
    data = dat,
    method = "rf",
    folds = 2,
    seed = 1)
} else {
  message("Package 'randomForest' not available; skipping example.")
}
}
