% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCI.tuner.R
\name{CCI.pretuner}
\alias{CCI.pretuner}
\alias{tuner}
\title{CCI tuner function for CCI test}
\usage{
CCI.pretuner(
  formula,
  data,
  method = "rf",
  folds = 10,
  tune_length = 10,
  seed = 1984,
  metric = NULL,
  ...
)
}
\arguments{
\item{formula}{Model formula or a DAGitty object specifying the relationship between dependent and independent variables.}

\item{data}{A data frame containing the variables specified in the formula.}

\item{method}{Character. Specifies the machine learning method to use. Supported methods are random forest "rf", extreme gradient boosting "xgboost", neural-net "nnet, Gaussian Process Regression "gpr" and Support Vector Machine "svm".}

\item{folds}{Integer. The number of folds for cross-validation during the tuning process. Default is 10.}

\item{tune_length}{Integer. The number of parameter combinations to try during the tuning process. Default is 10.}

\item{...}{Additional arguments to pass to the \code{CCI.tuner} function.}

\item{random_grid}{Logical. If TRUE, a random grid search is performed. If FALSE, a full grid search is performed. Default is TRUE.}
}
\value{
Returns tuned parameters values for the predictive function
}
\description{
The \code{CCI.tuner} function performs a grid search over parameters for a conditional independence test using machine learning model supported by CCI.test. The tuner use the caret package for tuning.
}
\seealso{
\code{\link{CCI.test} \link{perm.test}}, \code{\link{print.summary.CCI}}, \code{\link{plot.CCI}}, \code{\link{QQplot}}
}
