% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCI.test.R
\name{CCI.test}
\alias{CCI.test}
\title{Wrapper function for performing the CCI test}
\usage{
CCI.test(
  formula = NA,
  data,
  plot = TRUE,
  p = 0.825,
  nperm = 500,
  dag = NA,
  dag_n = 1,
  data_type = "continuous",
  method = "rf",
  parametric = FALSE,
  seed = NULL,
  metricfunc = NULL,
  mlfunc = NULL,
  tail = NA,
  ...
)
}
\arguments{
\item{formula}{Model formula or DAGitty object}

\item{data}{Data frame}

\item{plot}{Logical, indicating if a plot of null distribution with test statistic should be generated (default: TRUE)}

\item{p}{Proportion of data used for training}

\item{nperm}{Number of permutations}

\item{dag}{DAGitty object for conditional independence testing}

\item{dag_n}{Which test to perform if using a DAGitty object}

\item{data_type}{Type of data: "continuous", "binary", or "categorical"}

\item{method}{Method for modeling: "lm", "xgboost", "rf", etc.}

\item{parametric}{Logical, indicating if parametric p-value should be computed}

\item{seed}{Set seed value for reproduce results}

\item{metricfunc}{Custom made performance metric function}

\item{mlfunc}{Custom made machine learning model for prediction}

\item{tail}{Indicator, whether to calculate left or right tailed p-values, depends on the performance metric being used. Only use with metricfunc or MLfunc.}

\item{...}{Additional arguments to pass to \code{perm.test}}
}
\value{
Invisibly returns the result of \code{perm.test}
}
\description{
Wrapper function for performing the CCI test
}
\examples{
set.seed(123)
# Basic use
data <- data.frame(x1 = rnorm(100), x2 = rnorm(100), y = rnorm(100))
CCI.test(y ~ x1 | x2, data = data, nperm = 1000)
}
