% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCI.test.R
\name{CCI.test}
\alias{CCI.test}
\title{Wrapper function for performing the CCI test}
\usage{
CCI.test(
  formula = NA,
  data,
  plot = TRUE,
  p = 0.825,
  nperm = 500,
  dag = NA,
  dag_n = 1,
  data_type = "continuous",
  method = "rf",
  parametric = FALSE,
  seed = NULL,
  ...
)
}
\arguments{
\item{formula}{Model formula or DAGitty object}

\item{data}{Data frame}

\item{plot}{Logical, indicating if a plot of null distribution with test statistic should be generated (default: TRUE)}

\item{p}{Proportion of data used for training}

\item{nperm}{Number of permutations}

\item{dag}{DAGitty object for conditional independence testing}

\item{dag_n}{Which test to perform if using a DAGitty object}

\item{data_type}{Type of data: "continuous", "binary", or "categorical"}

\item{method}{Method for modeling: "lm", "xgboost", "rf", etc.}

\item{parametric}{Logical, indicating if parametric p-value should be computed}

\item{...}{Additional arguments to pass to \code{perm.test}}
}
\value{
Invisibly returns the result of \code{perm.test}
}
\description{
Wrapper function for performing the CCI test
}
