% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_formula}
\alias{check_formula}
\title{Check the logic of the statement}
\usage{
check_formula(formula, data)
}
\arguments{
\item{formula_str}{Formula string}
}
\description{
Check the logic of the statement
}
