% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_formula}
\alias{clean_formula}
\title{#' Clean formula string
#'
#' @param formula Formula
#'
#' @return Cleaned formula in the right format
#' @export}
\usage{
clean_formula(formula)
}
\description{
#' Clean formula string
#'
#' @param formula Formula
#'
#' @return Cleaned formula in the right format
#' @export
}
