% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_formula}
\alias{clean_formula}
\title{Clean formula string}
\usage{
clean_formula(formula_str)
}
\arguments{
\item{formula_str}{Formula string}
}
\value{
Cleaned formula string
}
\description{
Clean formula string
}
