% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ExponentialNoise}
\alias{ExponentialNoise}
\title{Generate Data with Exponential Noise}
\usage{
ExponentialNoise(N, rate_param = 1)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{rate_param}{Numeric. Rate parameter for the exponential distribution. Default is 1.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Adds exponential noise to a nonlinear combination of Z1 and Z2.
}
\examples{
head(ExponentialNoise(100))

}
