% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pvalues}
\alias{get_pvalues}
\title{P-value calculation based on null distribution and test statistic}
\usage{
get_pvalues(
  dist,
  test_statistic,
  parametric = FALSE,
  tail = c("left", "right")
)
}
\arguments{
\item{dist}{A vector representing the null-distribution}

\item{test_statistic}{The test statistic}

\item{parametric}{Logical, whether the perform to calculate parametric p-values}

\item{tail}{Indicator, whether to calculate left or right tailed p-values, depends on the performance metric being used}
}
\value{
P-value
}
\description{
P-value calculation based on null distribution and test statistic
}
