% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{HardCase}
\alias{HardCase}
\title{Generate Hard Case Data with Two Z Variables}
\usage{
HardCase(N)
}
\arguments{
\item{N}{Integer. Sample size.}
}
\value{
A data frame with columns X, Y, Z1, and Z2.
}
\description{
Generates data with a hard case scenario where X and Y are influenced by two Z variables in a nonlinear manner.
}
\examples{
head(HardCase(100))

}
