% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_strata_from_categorical_Z}
\alias{make_strata_from_categorical_Z}
\title{Create strata from the categorical subset of Z}
\usage{
make_strata_from_categorical_Z(sub_data, Z, allow_character = TRUE)
}
\arguments{
\item{sub_data}{data.frame containing Z columns.}

\item{Z}{character vector of Z column names.}

\item{allow_character}{logical; treat character as categorical. Default TRUE.}
}
\value{
A factor defining strata.
}
\description{
Uses interaction() on the categorical Z columns only.
}
