% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{multinom_wrapper}
\alias{multinom_wrapper}
\title{Wrapper function for multinomial regression model training and evaluation}
\usage{
multinom_wrapper(
  formula,
  data,
  train_indices,
  test_indices,
  metricfunc = NULL,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}

\item{train_indices}{Indices for training data}

\item{test_indices}{Indices for testing data}

\item{metricfunc}{A user-specified metric function which has the arguments data, model, and test_indices, and returns a numeric value}

\item{...}{Additional arguments passed to \code{nnet::multinom}}
}
\value{
Performance metric (Kappa for classification tasks)
}
\description{
Wrapper function for multinomial regression model training and evaluation
}
