% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{NonLinearCategorization}
\alias{NonLinearCategorization}
\title{Generate Nonlinear Categorical Data (Univariate)}
\usage{
NonLinearCategorization(N, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{d}{Numeric. Dependency strength. Default is 0.}
}
\value{
A data frame with columns Z, X, and Y.
}
\description{
Generates a dataset with a single Z influencing categorical X and Y.
}
