% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{NonLinearData}
\alias{NonLinearData}
\title{Generate Nonlinear Categorical Data (Bivariate)}
\usage{
NonLinearData(N, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{d}{Numeric. Dependency strength. Default is 0.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Creates categorical X and Y variables based on sinusoidal and cosine functions of Z1 and Z2.
}
