% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NonLinNormal}
\alias{NonLinNormal}
\title{Example dataset: NonLinNormal}
\format{
A data frame with 500 rows and 4 variables:
\describe{
\item{X}{Numeric vector}
\item{Y}{Numeric vector}
\item{Z1}{Conditioning variable}
\item{Z2}{Conditioning variable}
}
}
\source{
Simulated data.
}
\usage{
NonLinNormal
}
\description{
A dataset containing simulated data from a non-linear transformation of a multivariate normal distribution.
}
\keyword{datasets}
