% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{NormalData}
\alias{NormalData}
\title{Example dataset: NormalData}
\format{
A data frame with 400 rows and 4 variables:
\describe{
\item{X}{Numeric vector}
\item{Y}{Numeric vector}
\item{Z1}{Conditioning variable}
\item{Z2}{Conditioning variable}
}
@source Simulated data.
}
\usage{
NormalData
}
\description{
A dataset containing simulated data from a multivariate normal distribution.
}
\keyword{datasets}
