% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.test.R
\name{perm.test}
\alias{perm.test}
\title{Permutation test for conditional independence}
\usage{
perm.test(formula, data, MLfunc = lm, nperm = 1000, dag_n = 1, ...)
}
\arguments{
\item{formula}{Model formula or DAGitty object}

\item{data}{Data frame}

\item{MLfunc}{Model fitting function similar to \code{lm} (default)}

\item{nperm}{Number of permutations}

\item{dag_n}{Which test to perform if using a DAGitty object}

\item{...}{Additional arguments to pass to \code{MLfunc}}
}
\value{
An object of class 'CCI' containing a null distribution,
observed value, p-values, the ML model, and the data.
}
\description{
Permutation test for conditional independence
}
\examples{
dat <- data.frame(x1 = rnorm(100), x2 = rnorm(100), y = rnorm(100))
perm.test(y ~ x1 + x2, data = dat)
}
\seealso{
\code{\link{print.CCI}}, \code{\link{summary.CCI}},
\code{\link{plot.CCI}}, \code{\link{QQplot}}
}
