% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.CCI}
\alias{plot.CCI}
\title{Plot of null distribution and observed value}
\usage{
\method{plot}{CCI}(x, y = x$bin.width, overlay = c("none", "norm", "density"), ...)
}
\arguments{
\item{x}{Object of class 'CCI'}

\item{y}{Bin width for ggplot2's histogram}

\item{overlay}{Overlay curve options: 'none', 'norm', 'density'}

\item{...}{Additional arguments to be passed to ggplot2}
}
\value{
A plot of the null distribution and observed value in ggplot2 format.
}
\description{
Plot of null distribution and observed value
}
\examples{
dat <- data.frame(x1 = rnorm(100), x2 = rnorm(100), y = rnorm(100))
cci <- perm.test(y ~ x1 + x2, data = dat)
plot(cci)
}
\seealso{
\code{\link{print.CCI}}, \code{\link{summary.CCI}},
\code{\link{perm.test}}, \code{\link{QQplot}}
}
