% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.CCI}
\alias{plot.CCI}
\title{Plot for CCI testing}
\usage{
\method{plot}{CCI}(x, ...)
}
\arguments{
\item{x}{Object of class 'CCI'}

\item{...}{Additional arguments to ggplot2}
}
\value{
A plot of the null distribution and the test statistic in ggplot2 format.
}
\description{
Plot for CCI testing
}
\examples{
dat <- data.frame(x1 = rnorm(100), x2 = rnorm(100), y = rnorm(100))
cci <- perm.test(y ~ x1 + x2, data = dat)
plot(cci)
}
\seealso{
\code{\link{print.CCI}}, \code{\link{summary.CCI}},
\code{\link{plot.CCI}}, \code{\link{perm.test}}
}
