% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{PoissonNoise}
\alias{PoissonNoise}
\title{Example dataset: PoissonNoise}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
\item{X}{Numeric vector}
\item{Y}{Numeric vector}
\item{Z1}{Conditioning variable}
\item{Z2}{Conditioning variable}
}
}
\source{
Simulated data.
}
\usage{
PoissonNoise
}
\description{
A dataset containing simulated data from a Poisson distribution.
}
\keyword{datasets}
