% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{PolyData}
\alias{PolyData}
\title{Generate Categorical Polynomial Data}
\usage{
PolyData(N, d = 0)
}
\arguments{
\item{N}{Integer. Sample size.}

\item{d}{Numeric. Dependency strength. Default is 0.}
}
\value{
A data frame with columns Z1, Z2, X, and Y.
}
\description{
Generates X and Y categories based on polynomial combinations of Z1 and Z2.
}
