% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQplot.R
\name{QQplot}
\alias{QQplot}
\title{QQ-plot for multiple testing in CCI}
\usage{
QQplot(object, n, ...)
}
\arguments{
\item{object}{Object of class 'CCI'}

\item{n}{Number of samplings}

\item{...}{Additional arguments to ggplot2}
}
\value{
A QQ-plot of the p-values in ggplot2 format.
}
\description{
QQ-plot for multiple testing in CCI
}
\examples{
dat <- data.frame(x1 = rnorm(100), x2 = rnorm(100), y = rnorm(100))
cci <- perm.test(y ~ x1 + x2, data = dat)
QQplot(cci, 1000)
}
\seealso{
\code{\link{print.CCI}}, \code{\link{summary.CCI}},
\code{\link{plot.CCI}}, \code{\link{perm.test}}
}
