% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQplot.R
\name{QQplot}
\alias{QQplot}
\title{QQ-plot for multiple testing in CCI}
\usage{
QQplot(object, ...)
}
\arguments{
\item{object}{Object of class 'CCI'}

\item{...}{Additional arguments to pass to the \code{test.gen} function.}
}
\value{
A QQ-plot of the p-values in ggplot2 format.
}
\description{
QQ-plot for multiple testing in CCI
}
\examples{
dat <- data.frame(x1 = rnorm(200), x2 = rnorm(200), y = rnorm(200))
cci <- CCI.test(y ~ x1 | x2,
data = dat,
nperm = 50,
interaction = FALSE)
QQplot(cci)
}
\seealso{
\code{\link{print.CCI}}, \code{\link{summary.CCI}},
\code{\link{plot.CCI}}, \code{\link{perm.test}}
}
