% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ranger_wrapper}
\alias{ranger_wrapper}
\title{Wrapper function for Ranger model training and evaluation}
\usage{
ranger_wrapper(
  formula,
  data,
  train_indices,
  test_indices,
  probability = FALSE,
  num.trees,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}

\item{train_indices}{Indices for training data}

\item{test_indices}{Indices for training data}

\item{probability}{Grow a probability forest for binary or categorical outcomes}

\item{num.trees}{Number of boosting rounds}

\item{...}{Additional arguments passed to ranger}

\item{iteration}{Current iteration index}
}
\value{
Performance metric (RMSE for continuous, Kappa for binary, Kappa for categorical)
}
\description{
Wrapper function for Ranger model training and evaluation
}
