% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{print.CCI}
\alias{print.CCI}
\alias{reports}
\alias{summary.CCI}
\alias{print.summary.CCI}
\title{Print and summary methods for the CCI class}
\usage{
\method{print}{CCI}(x, ...)

\method{summary}{CCI}(object, ...)

\method{print}{summary.CCI}(x, ...)
}
\arguments{
\item{x}{Object of class 'CCI'}

\item{...}{Additional arguments to print/summary}

\item{object}{Object of class 'CCI'}
}
\value{
The print methods have no return value, the summary methods return
an object of class 'summary.CCI'.
}
\description{
Print and summary methods for the CCI class
}
