% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.gen.R
\name{test.gen}
\alias{test.gen}
\title{Generate the test statistic or null distribution using permutation}
\usage{
test.gen(
  Y,
  X,
  Z,
  data,
  data_type = "continuous",
  method = "rf",
  nperm = 100,
  p = 0.85,
  N = nrow(data),
  poly = TRUE,
  degree = 3,
  nrounds = 120,
  family,
  objective = "reg:squarederror",
  num_class = NULL,
  probability = FALSE,
  permutation = FALSE,
  metricfunc = NULL,
  mlfunc = NULL,
  seed = NULL,
  ...
)
}
\arguments{
\item{Y}{Dependent variable}

\item{X}{Independent variable to be permuted}

\item{Z}{Conditioning variables}

\item{data}{Data frame}

\item{data_type}{Type of data: "continuous", "binary", or "multinomial"}

\item{method}{Method for modeling: "lm", "xgboost", "rf"}

\item{nperm}{Number of generated samples}

\item{p}{Proportion of data used for training}

\item{N}{Number of observations}

\item{poly}{Logical, whether to include polynomial terms}

\item{degree}{Degree of polynomial terms}

\item{nrounds}{Number of rounds (trees) for xgboost and ranger}

\item{family}{Family for glm}

\item{objective}{Objective function for xgboost}

\item{probability}{Logical, whether the ranger_wrapper should do classification}

\item{permutation}{Logical, whether the perform permutation to generat a null distribution}

\item{mlfunc}{custom ML function provided by the user, the function must have the arguments: formula, data, train_indices, test_indices and ..., and return a single value performance metric}

\item{...}{additional arguments to pass to the machine learning wrapper functions \code{glm_wrapper}, \code{multinom_wrapper}, \code{xgboost_wrapper}, \code{ranger_wrapper}. Or to a custom build wrapper function.}
}
\value{
A list containing the test distribution
}
\description{
Generate the test statistic or null distribution using permutation
}
