% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.gen.R
\name{test.gen}
\alias{test.gen}
\title{Generate the Test Statistic or Null Distribution Using Permutation}
\usage{
test.gen(
  Y,
  X,
  Z,
  data,
  data_type = "continuous",
  method = "rf",
  nperm = 100,
  p = 0.85,
  N = nrow(data),
  poly = TRUE,
  degree = 3,
  nrounds = 120,
  family,
  num_class = NULL,
  nthread = 1,
  permutation = FALSE,
  metricfunc = NULL,
  mlfunc = NULL,
  ...
)
}
\arguments{
\item{Y}{Character. The name of the dependent (response) variable in the data.}

\item{X}{Character. The name of the independent (predictor) variable to be permuted.}

\item{Z}{Character vector. The names of the conditioning variables to be included in the model.}

\item{data}{Data frame. The data containing the variables used in the analysis.}

\item{data_type}{Character. The type of data of the Y parameter: can be "continuous", "binary", or "multinomial".}

\item{method}{Character. The modeling method to be used. Options include "lm" for linear models, "xgboost" for gradient boosting, or "rf" for random forests.}

\item{nperm}{Integer. The number of generated samples or permutations. Default is 100.}

\item{p}{Numeric. The proportion of the data to be used for training. The remaining data will be used for testing. Default is 0.85.}

\item{N}{Integer. The total number of observations in the data. Default is the number of rows in the data frame.}

\item{poly}{Logical. Whether to include polynomial terms of the conditioning variables. Default is TRUE.}

\item{degree}{Integer. The degree of polynomial terms to be included if \code{poly} is TRUE. Default is 3.}

\item{nrounds}{Integer. The number of rounds (trees) for methods like xgboost and ranger. Default is 120.}

\item{family}{Family object. The family parameter for generalized linear models (e.g., \code{gaussian()} for linear regression).}

\item{num_class}{Integer. The number of classes for multinomial classification (used in xgboost with \code{objective = "multi:softprob"}). Default is NULL.}

\item{permutation}{Logical. Whether to perform permutation to generate a null distribution. Default is FALSE.}

\item{metricfunc}{Function. A custom metric function provided by the user. The function must take arguments: \code{data}, \code{model}, \code{test_indices}, and \code{test_matrix}, and return a single value performance metric. Default is NULL.}

\item{mlfunc}{Function. A custom machine learning function provided by the user. The function must have the arguments: \code{formula}, \code{data}, \code{train_indices}, \code{test_indices}, and \code{...}, and return a single value performance metric. Default is NULL.}

\item{...}{Additional arguments to pass to the machine learning wrapper functions \code{glm_wrapper}, \code{multinom_wrapper}, \code{xgboost_wrapper}, \code{ranger_wrapper}, or to a custom-built wrapper function.}
}
\value{
A list containing the test distribution.
}
\description{
This function generates the test statistic or a null distribution through permutation for conditional independence testing.
It supports various machine learning methods, including linear models, random forests, and gradient boosting, and allows for custom metric functions and model fitting functions.
}
\examples{
set.seed(123)
data <- data.frame(x1 = rnorm(100),
x2 = rnorm(100),
x3 = rnorm(100),
x4 = rnorm(100),
y = rnorm(100))
result <- test.gen(Y = "y", X = "x1", Z = c("x2", "x3", "x4"), data = data)
}
