% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrapper_svm}
\alias{wrapper_svm}
\title{Wrapper function for SVM model training and evaluation}
\usage{
wrapper_svm(
  formula,
  data,
  train_indices,
  test_indices,
  data_type,
  metricfunc = NULL,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}

\item{train_indices}{Indices for training data}

\item{test_indices}{Indices for testing data}

\item{data_type}{Type of data ("continuous", "binary", or "categorical")}

\item{metricfunc}{Optional user-defined function to calculate a custom performance metric.}

\item{...}{Additional arguments passed to e1071::svm}
}
\value{
Performance metric (RMSE for continuous, Kappa for classification)
}
\description{
Wrapper function for SVM model training and evaluation
}
