% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{wrapper_xgboost}
\alias{wrapper_xgboost}
\title{Wrapper function for XGBoost model training and evaluation}
\usage{
wrapper_xgboost(
  formula,
  data,
  train_indices,
  test_indices,
  data_type,
  nrounds,
  num_class = NULL,
  metricfunc = NULL,
  nthread = 1,
  ...
)
}
\arguments{
\item{formula}{Model formula}

\item{data}{Data frame}

\item{train_indices}{Indices for training data}

\item{test_indices}{Indices for training data}

\item{data_type}{Type of data (continuous, binary or categorical)}

\item{nrounds}{Number of boosting rounds}

\item{num_class}{Number of categorical classes}

\item{metricfunc}{A user specific metric function which have the arguments data, model test_indices and test_matrix and returns a numeric value}

\item{nthread}{Integer. Number of threads to use for parallel computation during model training in XGBoost. Default is 1.}

\item{...}{Additional arguments passed to xgb.train}
}
\value{
Performance metric
}
\description{
Wrapper function for XGBoost model training and evaluation
}
